Using Python Socket Programming
Here, I Created a GUI chat program written in python2.7. It is designed to not need a central server and so work very well on LANs that are not connected to the internet. One user starts the program in server mode, listening on a port of their choosing, and then other users will use client mode to connect to their machine.